backend/app/models.py
from app import db

class Song(db.Model):
    id = db.Column(db.String(50), primary_key=True)
    title = db.Column(db.String(100), nullable=False)
    index = db.Column(db.String(50))
    danceability = db.Column(db.Float)
    energy = db.Column(db.Float)
    key = db.Column(db.Integer)
    loudness = db.Column(db.Float)
    mode = db.Column(db.Integer)
    acousticness = db.Column(db.Float)
    instrumentalness = db.Column(db.Float)
    liveness = db.Column(db.Float)
    valence = db.Column(db.Float)
    tempo = db.Column(db.Float)
    duration_ms = db.Column(db.Integer)
    time_signature = db.Column(db.Integer)
    num_bars = db.Column(db.Integer)
    num_sections = db.Column(db.Integer)
    num_segments = db.Column(db.Integer)
    # Changed from 'class' to 'class_' to handle an error python throws for using the restricted keyword
    class_ = db.Column(db.Integer)  
    rating = db.Column(db.Integer, default=0)

    def to_dict(self):
        return {column.name: getattr(self, column.name) for column in self.__table__.columns}